requirements.txt
pytest
falcon
jinja2
waitress
python-dotenv
mysql-connector-python
bcrypt
tzdata